+++ src/dialog/windows/SDL_windowsdialog.c
@@ -1028,7 +1028,7 @@ void windows_ShowFileDialog(void *ptr)
             const char *opts[1] = { NULL };
             callback(userdata, opts, getFilterIndex(dialog.nFilterIndex));
         } else {
-            SDL_SetError("Windows error, CommDlgExtendedError: %ld", pCommDlgExtendedError());
+            SDL_SetError("Windows error, CommDlgExtendedError: %ld", (long)pCommDlgExtendedError());
             callback(userdata, NULL, -1);
         }
     }
